requirements.txt
numpy
torch==1.10
transformers
nltk
pandas
pyyaml
torchtext
emoji
URLExtract